product backend

+++ backend/controllers/categoryController.js
@@ -10,6 +10,7 @@
 import asyncHandler from 'express-async-handler';
 import Category from "../models/categoryModel.js";
 import SubCategory from '../models/subCategoryModel.js';
+import Product from '../models/productModel.js';
 import {uploadCategory} from '../uploadFile.js';
 import multer from 'multer';
 import fs from 'fs';
@@ -152,10 +153,10 @@ const deleteCategory = asyncHandler(async (req, res) => {
         }
 
         // Check if any products reference this category
-        // const products = await Product.find({ proCategoryId: categoryID });
-        // if (products.length > 0) {
-        //     return res.status(400).json({ success: false, message: "Cannot delete category. Products are referencing it." });
-        // }
+        const products = await Product.find({ proCategoryId: categoryID });
+        if (products.length > 0) {
+            return res.status(400).json({ success: false, message: "Cannot delete category. Products are referencing it." });
+        }
 
         // If no subcategories or products are referencing the category, proceed with deletion
         const category = await Category.findByIdAndDelete(categoryID);
